Moving subfilters refresh/load logic from tabs fragment to root bottomsheet fragment.
